requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9
email-validator==2.2.0
pytest==8.3.3
ruff==0.6.8